// sdk/nodejs/sdk/src/api/models/AuthUser.ts
/* tslint:disable */
/* eslint-disable */
/**
 * zrok
 * zrok client access
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuthUser
 */
export interface AuthUser {
    /**
     * 
     * @type {string}
     * @memberof AuthUser
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthUser
     */
    password?: string;
}

/**
 * Check if a given object implements the AuthUser interface.
 */
export function instanceOfAuthUser(value: object): value is AuthUser {
    return true;
}

export function AuthUserFromJSON(json: any): AuthUser {
    return AuthUserFromJSONTyped(json, false);
}

export function AuthUserFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthUser {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'] == null ? undefined : json['username'],
        'password': json['password'] == null ? undefined : json['password'],
    };
}

export function AuthUserToJSON(json: any): AuthUser {
    return AuthUserToJSONTyped(json, false);
}

export function AuthUserToJSONTyped(value?: AuthUser | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'password': value['password'],
    };
}