// sdk/nodejs/sdk/src/api/models/UnshareRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * zrok
 * zrok client access
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UnshareRequest
 */
export interface UnshareRequest {
    /**
     * 
     * @type {string}
     * @memberof UnshareRequest
     */
    envZId?: string;
    /**
     * 
     * @type {string}
     * @memberof UnshareRequest
     */
    shareToken?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UnshareRequest
     */
    reserved?: boolean;
}

/**
 * Check if a given object implements the UnshareRequest interface.
 */
export function instanceOfUnshareRequest(value: object): value is UnshareRequest {
    return true;
}

export function UnshareRequestFromJSON(json: any): UnshareRequest {
    return UnshareRequestFromJSONTyped(json, false);
}

export function UnshareRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UnshareRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'envZId': json['envZId'] == null ? undefined : json['envZId'],
        'shareToken': json['shareToken'] == null ? undefined : json['shareToken'],
        'reserved': json['reserved'] == null ? undefined : json['reserved'],
    };
}

export function UnshareRequestToJSON(json: any): UnshareRequest {
    return UnshareRequestToJSONTyped(json, false);
}

export function UnshareRequestToJSONTyped(value?: UnshareRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'envZId': value['envZId'],
        'shareToken': value['shareToken'],
        'reserved': value['reserved'],
    };
}